finish off implementation of initio (for now) and begin web server work

+++ initio/servos.go
@@ -13,6 +13,7 @@ const (
 
 var servosActive = false
 
+// set a servo to a certain angle
 func SetServo(servo int, degrees int) {
 	if !servosActive {
 		StartServos()
@@ -36,7 +37,12 @@ func StartServos() {
 	}
 }
 
+// start servod, the servo daemon
 func startServod() error {
+	if servosActive {
+		// already open
+		return nil
+	}
 
 	// run the command ./servod --pcm --idle-timeout=20000 --p1pins="18,22"
 	cmd := exec.Command("sudo", "sh", "-c", "servod --pcm --idle-timeout=20000 --p1pins=\"18,22\"")
@@ -55,7 +61,13 @@ func startServod() error {
 	return err
 }
 
+// stop servod, the servo daemon
 func stopServod() error {
+	if !servosActive {
+		// don't close something that isn't there
+		return nil
+	}
+
 	cmd := exec.Command("sudo", "sh", "-c", "killall servod")
 	err := cmd.Run()
 
@@ -70,6 +82,7 @@ func stopServod() error {
 	return err
 }
 
+// apply servo change to the pin
 func pinServod(pin int, degrees int) {
 	pinString := fmt.Sprintf("%d=%d\n", pin, 50+((90-degrees)*200/180))
 	fmt.Println(pinString)
